Canvis no confirmats
Importació alumnes saga i esfera
Mitjançant paràmetres a Paràmetres Saga:
mantenirNom si 'True' manté el nom de pila de l'alumne encara que sigui
diferent a la importació.
mantenirLlistes si 'True' manté els alumnes a les llistes encara que
hagin canviat de grup classe a la importació.

Incidències tutor
Permet visualitzar totes les incidències dels alumnes de la tutoria
encara que les hagin posat altres professors.

+++ aula/apps/extSaga/sincronitzaSaga.py
@@ -215,6 +215,12 @@ def sincronitza(f, user = None):
                 a.tutors_volen_rebre_correu           = alumneDadesAnteriors.tutors_volen_rebre_correu = False
                 a.foto = alumneDadesAnteriors.foto
 
+            # amorilla@xtec.cat
+            manteNom, _ = ParametreSaga.objects.get_or_create( nom_parametre = 'mantenirNom' )
+            
+            if manteNom.valor_parametre=='True':
+                a.nom=alumneDadesAnteriors.nom
+
         a.save()
         nivells.add(a.grup.curs.nivell)
     #
@@ -294,17 +300,20 @@ def sincronitza(f, user = None):
                 impartir__dia_impartir__gte = date.today(),
                 alumne__in = AlumnesDonatsDeBaixa ).delete()
 
-    #Treure'ls de les classes: els canvis de grup   #Todo: només si l'àmbit és grup.
-
-    ambit_no_es_el_grup = Q( impartir__horari__assignatura__tipus_assignatura__ambit_on_prendre_alumnes__in = [ 'C', 'N', 'I' ] )
-    ( ControlAssistencia
-      .objects
-      .filter( ambit_no_es_el_grup )
-      .filter( impartir__dia_impartir__gte = date.today() )
-      .filter( alumne__in = AlumnesCanviatsDeGrup )
-      .delete()
-     )
-
+    # amorilla@xtec.cat
+    manteLlista, _ = ParametreSaga.objects.get_or_create( nom_parametre = 'mantenirLlistes' )
+            
+    if manteLlista.valor_parametre!='True':
+        #Treure'ls de les classes: els canvis de grup   #Todo: només si l'àmbit és grup.
+    
+        ambit_no_es_el_grup = Q( impartir__horari__assignatura__tipus_assignatura__ambit_on_prendre_alumnes__in = [ 'C', 'N', 'I' ] )
+        ( ControlAssistencia
+          .objects
+          .filter( ambit_no_es_el_grup )
+          .filter( impartir__dia_impartir__gte = date.today() )
+          .filter( alumne__in = AlumnesCanviatsDeGrup )
+          .delete()
+         )
 
     #Altes: posar-ho als controls d'assistència de les classes (?????????)
